# Refactor lib/cli/serve.js and lib/cli/watch.js to fix some major performance issues in development