--- a/lib/cli/serve.js
+++ b/lib/cli/serve.js
@@ -15,20 +15,19 @@ var getOutputFolder = require('./getOutputFolder');
 var Server = require('./server');
 var watch = require('./watch');
 
-var server, lrServer, lrPath;
+var server, lrServer, lrFilepaths = [];
 
 function waitForCtrlC() {
     var d = Promise.defer();
 
-    process.on('SIGINT', function() {
+    process.on('SIGINT', function () {
         d.resolve();
     });
 
     return d.promise;
 }
 
-
-function generateBook(args, kwargs) {
+function startServer(args, kwargs) {
     var port = kwargs.port;
     var outputFolder = getOutputFolder(args);
     var book = getBook(args, kwargs);
@@ -38,59 +37,78 @@ function generateBook(args, kwargs) {
     var hasWatch = kwargs['watch'];
     var hasLiveReloading = kwargs['live'];
     var hasOpen = kwargs['open'];
+    var regenerationInProgress = false;
+    var needsRegeneration = false;
 
-    // Stop server if running
-    if (server.isRunning()) console.log('Stopping server');
-
-    return server.stop()
-    .then(function() {
+    function generateBook() {
         return Parse.parseBook(book)
-        .then(function(resultBook) {
+            .then(function (resultBook) {
+                if (hasLiveReloading) {
+                    // Enable livereload plugin
+                    var config = resultBook.getConfig();
+                    config = ConfigModifier.addPlugin(config, 'livereload');
+                    resultBook = resultBook.set('config', config);
+                }
+                return Output.generate(Generator, resultBook, {
+                    root: outputFolder
+                });
+            });
+    }
+
+    // Also triggers livereload update
+    function regenerateBook() {
+        regenerationInProgress = true;
+        return generateBook(args, kwargs).then(function () {
             if (hasLiveReloading) {
-                // Enable livereload plugin
-                var config = resultBook.getConfig();
-                config = ConfigModifier.addPlugin(config, 'livereload');
-                resultBook = resultBook.set('config', config);
+                // trigger livereload
+                console.log('Triggering livereload for', lrFilepaths);
+                lrServer.changed({
+                    body: {
+                        files: [lrFilepaths]
+                    }
+                });
             }
 
-            return Output.generate(Generator, resultBook, {
-                root: outputFolder
-            });
+            if (needsRegeneration) {
+                needsRegeneration = false;
+                return regenerateBook();
+            }
+            regenerationInProgress = false;
+            lrFilepaths = [];
         });
-    })
-    .then(function() {
-        console.log();
-        console.log('Starting server ...');
-        return server.start(outputFolder, port);
-    })
-    .then(function() {
-        console.log('Serving book on http://localhost:'+port);
+    }
 
-        if (lrPath && hasLiveReloading) {
-            // trigger livereload
-            lrServer.changed({
-                body: {
-                    files: [lrPath]
+    console.log('Generating book ...');
+    return generateBook(args, kwargs).then(function () {
+        console.log('Starting server ...');
+        return server.start(outputFolder, port).then(function () {
+            console.log('Serving book on http://localhost:' + port);
+
+            if (hasOpen) {
+                open('http://localhost:' + port, browser);
+            }
+
+            if (!hasWatch) {
+                return waitForCtrlC();
+            }
+
+            watch(book.getRoot(), function (err, filepath) {
+                if (err) {
+                    console.log(err);
+                    return;
+                }
+                if (!lrFilepaths.includes(filepath)) {
+                    lrFilepaths.push(filepath);
+                }
+                if (regenerationInProgress) {
+                    console.log('Queuing another regeneration after change in', filepath);
+                    needsRegeneration = true;
+                } else {
+                    console.log('Regenerating book after change in', filepath);
+                    regenerateBook();
                 }
             });
-        }
-
-        if (hasOpen) {
-            open('http://localhost:'+port, browser);
-        }
-    })
-    .then(function() {
-        if (!hasWatch) {
             return waitForCtrlC();
-        }
-
-        return watch(book.getRoot())
-        .then(function(filepath) {
-            // set livereload path
-            lrPath = filepath;
-            console.log('Restart after change in file', filepath);
-            console.log('');
-            return generateBook(args, kwargs);
         });
     });
 }
@@ -132,28 +150,27 @@ module.exports = {
         options.log,
         options.format
     ],
-    exec: function(args, kwargs) {
+    exec: function (args, kwargs) {
         server = new Server();
         var hasWatch = kwargs['watch'];
         var hasLiveReloading = kwargs['live'];
 
         return Promise()
-        .then(function() {
-            if (!hasWatch || !hasLiveReloading) {
-                return;
-            }
+            .then(function () {
+                if (!hasWatch || !hasLiveReloading) {
+                    return;
+                }
 
-            lrServer = tinylr({});
-            return Promise.nfcall(lrServer.listen.bind(lrServer), kwargs.lrport)
-            .then(function() {
-                console.log('Live reload server started on port:', kwargs.lrport);
-                console.log('Press CTRL+C to quit ...');
-                console.log('');
-
+                lrServer = tinylr({});
+                return Promise.nfcall(lrServer.listen.bind(lrServer), kwargs.lrport)
+                    .then(function () {
+                        console.log('Live reload server started on port:', kwargs.lrport);
+                        console.log('Press CTRL+C to quit ...');
+                        console.log('');
+                    });
+            })
+            .then(function () {
+                return startServer(args, kwargs);
             });
-        })
-        .then(function() {
-            return generateBook(args, kwargs);
-        });
     }
 };