--- a/lib/cli/watch.js
+++ b/lib/cli/watch.js
@@ -10,8 +10,7 @@ var parsers = require('../parsers');
     @param {String} dir
     @return {Promise}
 */
-function watch(dir) {
-    var d = Promise.defer();
+function watch(dir, callback) {
     dir = path.resolve(dir);
 
     var toWatch = [
@@ -20,7 +19,7 @@ function watch(dir) {
 
     // Watch all parsable files
     parsers.extensions.forEach(function(ext) {
-        toWatch.push('**/*'+ext);
+        toWatch.push('**/*' + ext);
     });
 
     var watcher = chokidar.watch(toWatch, {
@@ -29,18 +28,12 @@ function watch(dir) {
         ignoreInitial: true
     });
 
-    watcher.once('all', function(e, filepath) {
-        watcher.close();
-
-        d.resolve(filepath);
+    watcher.on('all', function(e, filepath) {
+        callback(null, filepath);
     });
-    watcher.once('error', function(err) {
-        watcher.close();
-
-        d.reject(err);
+    watcher.on('error', function(err) {
+        callback(err);
     });
-
-    return d.promise;
 }
 
 module.exports = watch;