--- a/lib/cli/watch.js
+++ b/lib/cli/watch.js
@@ -24,7 +24,7 @@ function watch(dir, callback) {
 
     var watcher = chokidar.watch(toWatch, {
         cwd: dir,
-        ignored: '_book/**',
+        ignored: ['_book/**', 'node_modules/**'],
         ignoreInitial: true
     });