Fixes https://github.com/GitbookIO/gitbook/issues/1829

The main thing is that there's no need to stop and start the server. This step was really slow on my computer, and would sometimes take > 45 seconds after the first few changes.

The other thing is that I'm no longer returning a promise from `watch.js`. There's no need to run `chokidar.watch` every time a file changes, we can just set it up once. I pass in a callback that is called every time, so it doesn't need to do any teardown or setup. 

When a file changes, I'm now calling the `regenerateBook` function to regenerate the files. This function also triggers livereload when it's done, and it checks to see if any files have been changed while it was busy. If there any files change during the generation, then it will keep re-running `regenerateBook` until there are no more changes. (Before, it would miss some changes if you save a file, and then save again before the generation had finished.)

The console output now looks like this:


```
$ gitbook serve
Live reload server started on port: 35729
Press CTRL+C to quit ...

Generating book ...
info: 9 plugins are installed
info: 8 explicitly listed
info: loading plugin "ga"... OK
info: loading plugin "livereload"... OK
info: loading plugin "highlight"... OK
info: loading plugin "search"... OK
info: loading plugin "lunr"... OK
info: loading plugin "fontsettings"... OK
info: loading plugin "theme-api"... OK
info: loading plugin "theme-default"... OK
info: found 18 pages
info: found 58 asset files
warn: "this.generator" property is deprecated, use "this.output.name" instead
warn: "navigation" property is deprecated
warn: "book" property is deprecated, use "this" directly instead
warn: "options" property is deprecated, use config.get(key) instead
info: >> generation finished with success in 3.9s !
Starting server ...
Serving book on http://localhost:4000
Regenerating book after change in SUMMARY.md
Queuing another regeneration after change in foo.md
Queuing another regeneration after change in foo.md
Queuing another regeneration after change in SUMMARY.md
info: 9 plugins are installed
info: 8 explicitly listed
info: loading plugin "ga"... OK
info: loading plugin "livereload"... OK
info: loading plugin "highlight"... OK
info: loading plugin "search"... OK
info: loading plugin "lunr"... OK
info: loading plugin "fontsettings"... OK
info: loading plugin "theme-api"... OK
info: loading plugin "theme-default"... OK
info: found 18 pages
info: found 58 asset files
info: >> generation finished with success in 11.5s !
Triggering livereload for [ 'SUMMARY.md', 'foo.md' ]
info: 9 plugins are installed
info: 8 explicitly listed
info: loading plugin "ga"... OK
info: loading plugin "livereload"... OK
info: loading plugin "highlight"... OK
info: loading plugin "search"... OK
info: loading plugin "lunr"... OK
info: loading plugin "fontsettings"... OK
info: loading plugin "theme-api"... OK
info: loading plugin "theme-default"... OK
info: found 17 pages
info: found 59 asset files
info: >> generation finished with success in 2.9s !
Triggering livereload for [ 'SUMMARY.md', 'foo.md' ]
```